Continue porting to QGIS3
- Moved logic out of dialog
- removed custom QProgressBar
