+++ src/qgis_conefor/conefordialog.py
@@ -1,9 +1,7 @@
-import enum
 import os
 from typing import Optional
 from pathlib import Path
 
-from processing.tools.dataobjects import createContext
 import qgis.core
 import qgis.gui
 from qgis.PyQt import (
@@ -13,12 +11,9 @@ from qgis.PyQt import (
     uic,
 )
 
-from .processing.algorithms import coneforinputs
 from . import (
-    coneforinputsprocessor,
     schemas,
     tablemodel,
-    tasks,
 )
 from .utilities import (
     load_settings_key,
@@ -32,20 +27,9 @@ FORM_CLASS, _ = uic.loadUiType(str(UI_DIR / "conefor_dlg.ui"))
 
 class ConeforDialog(QtWidgets.QDialog, FORM_CLASS):
 
-    # processing algorithms
-    inputs_from_points_algorithm: qgis.core.QgsProcessingAlgorithm
-    inputs_from_polygons_algorithm: qgis.core.QgsProcessingAlgorithm
-    processing_context: qgis.core.QgsProcessingContext
-
-    # tasks
-    analyzer_task: qgis.core.QgsTask
-    processing_task: Optional[qgis.core.QgsTask]
-    processing_tasks: list[qgis.core.QgsTask]
-
     _layers: dict[qgis.core.QgsVectorLayer, list[str]]
     iface: qgis.gui.QgisInterface
     model: Optional[tablemodel.ProcessLayerTableModel]
-    processor: coneforinputsprocessor.InputsProcessor
 
     # UI controls
     add_row_btn: QtWidgets.QPushButton
@@ -55,7 +39,6 @@ class ConeforDialog(QtWidgets.QDialog, FORM_CLASS):
     edge_distance_rb: QtWidgets.QRadioButton
     layers_la: QtWidgets.QLabel
     lock_layers_chb: QtWidgets.QCheckBox
-    progressBar: QtWidgets.QProgressBar
     progress_la: QtWidgets.QLabel
     output_la: QtWidgets.QLabel
     output_dir_le: QtWidgets.QLineEdit
@@ -69,15 +52,14 @@ class ConeforDialog(QtWidgets.QDialog, FORM_CLASS):
         super(ConeforDialog, self).__init__(parent)
         self.setupUi(self)
         self.edge_distance_rb.setChecked(True)
-        self.processor = plugin_obj.processor
         self.model = None
         self.iface = plugin_obj.iface
         self.lock = QtCore.QReadWriteLock()
         self.change_ui_availability(False)
         self.buttonBox.button(self.buttonBox.Help).released.connect(self.show_help)
+        self.buttonBox.button(self.buttonBox.Cancel).released.connect(self.reject)
+        self.buttonBox.button(self.buttonBox.Ok).released.connect(self.accept)
         self.progress_la.setText('Analyzing layers...')
-        self.progressBar.setMinimum(0)
-        self.progressBar.setMaximum(0)
         self.use_selected_features_chb.setChecked(
             load_settings_key(
                 schemas.QgisConeforSettingsKey.USE_SELECTED,
@@ -88,21 +70,6 @@ class ConeforDialog(QtWidgets.QDialog, FORM_CLASS):
         self.use_selected_features_chb.stateChanged.connect(
             self.use_selected_features_toggled)
         self._layers = {}
-        self.processing_task = None
-        processing_registry = qgis.core.QgsApplication.processingRegistry()
-        self.inputs_from_points_algorithm = processing_registry.createAlgorithmById(
-            "conefor:inputsfrompoint")
-        self.inputs_from_polygons_algorithm = processing_registry.createAlgorithmById(
-            "conefor:inputsfrompolygon")
-        self.processing_context = createContext()
-        self.processing_tasks = []
-        task_manager = qgis.core.QgsApplication.taskManager()
-        self.analyzer_task = tasks.LayerAnalyzerTask(
-            description="analyze currently loaded layers",
-            layers_to_analyze=qgis.core.QgsProject.instance().mapLayers(),
-        )
-        self.analyzer_task.layers_analyzed.connect(self.finished_analyzing_layers)
-        task_manager.addTask(self.analyzer_task)
 
     def use_selected_features_toggled(self, state: int):
         save_settings_key(
@@ -139,7 +106,6 @@ class ConeforDialog(QtWidgets.QDialog, FORM_CLASS):
             self.model = tablemodel.ProcessLayerTableModel(
                 qgis_layers=self._layers,
                 initial_layers_to_process=selected_layers,
-                processor=self.processor,
                 dialog=self
             )
             self.tableView.setModel(self.model)
@@ -147,11 +113,9 @@ class ConeforDialog(QtWidgets.QDialog, FORM_CLASS):
             self.tableView.setItemDelegate(delegate)
             self.add_row_btn.released.connect(self.add_conefor_input)
             self.remove_row_btn.released.connect(self.remove_conefor_input)
-            self.buttonBox.button(self.buttonBox.Ok).released.connect(self.run_queries)
+            # self.buttonBox.button(self.buttonBox.Ok).released.connect(self.run_queries)
             self.output_dir_btn.released.connect(self.get_output_dir)
             self.lock_layers_chb.toggled.connect(self.toggle_lock_layers)
-            self.processor.progress_changed.connect(self.update_progress)
-            self.processor.update_info.connect(self.update_info)
             self.model.is_runnable_check.connect(self.toggle_run_button)
             if len(selected_layers) < 2:
                 self.remove_row_btn.setEnabled(False)
@@ -160,8 +124,6 @@ class ConeforDialog(QtWidgets.QDialog, FORM_CLASS):
                 schemas.QgisConeforSettingsKey.OUTPUT_DIR, default_to=str(Path.home()))
             self.output_dir_le.setText(output_dir)
             self.create_distances_files_chb.setChecked(False)
-            self.progressBar.setValue(self.processor.global_progress)
-            self.update_info("")
         else:
             self.change_ui_availability(False)
             self.progress_la.setText(
@@ -171,16 +133,10 @@ class ConeforDialog(QtWidgets.QDialog, FORM_CLASS):
             palette = QtGui.QPalette()
             palette.setColor(QtGui.QPalette.Foreground, QtCore.Qt.GlobalColor.red)
             self.progress_la.setPalette(palette)
-        self.reset_progress_bar()
 
     def toggle_lock_layers(self, lock):
         index = self.model.index(0, 0)
         self.tableView.setFocus()
-
-    def reset_progress_bar(self):
-        self.progressBar.setMinimum(0)
-        self.progressBar.setMaximum(100)
-        self.progressBar.setValue(0)
 
     def show_help(self):
         QtGui.QDesktopServices.openUrl(
@@ -215,177 +171,6 @@ class ConeforDialog(QtWidgets.QDialog, FORM_CLASS):
         save_settings_key(schemas.QgisConeforSettingsKey.OUTPUT_DIR, final_dir)
         self.output_dir_le.setText(final_dir)
 
-    def get_conefor_input_parameters(
-            self,
-            data_item: schemas.TableModelItem,
-            load_to_canvas: bool,
-            default_node_identifier_name: Optional[str] = None,
-            default_node_attribute_name: Optional[str] = None,
-            # default_process_centroid_distance: Optional[bool] = None,
-            # default_process_edge_distance: Optional[bool] = None,
-    ) -> schemas.ConeforInputParameters:
-        node_identifier_name = (
-                default_node_identifier_name or data_item.id_attribute_field_name)
-        if node_identifier_name != schemas.NONE_LABEL:
-            node_attribute_name = (
-                    default_node_attribute_name or data_item.attribute_field_name)
-            process_edge_distance = self.edge_distance_rb.isChecked()
-            layer_name = data_item.layer.name()
-            return schemas.ConeforInputParameters(
-                layer=data_item.layer,
-                id_attribute_field_name=(
-                    node_identifier_name
-                    if node_identifier_name != schemas.AUTOGENERATE_NODE_ID_LABEL
-                    else None
-                ),
-                attribute_field_name=(
-                    node_attribute_name if node_attribute_name != schemas.GENERATE_FROM_AREA_LABEL
-                    else None
-                ),
-                connections_method=(
-                    schemas.NodeConnectionType.EDGE_DISTANCE if process_edge_distance
-                    else schemas.NodeConnectionType.CENTROID_DISTANCE
-                )
-                # attribute_file_name=(
-                #     f"nodes_{node_attribute_name}_{layer_name}"
-                #     if node_attribute_name else None
-                # ),
-                # area_file_name=(
-                #     f"nodes_calculated_area_{layer_name}"
-                #     if process_area else None
-                # ),
-                # centroid_file_name=(
-                #     f"distances_centroids_{layer_name}"
-                #     if process_centroid_distance else None
-                # ),
-                # edge_file_name=(
-                #     f"distances_edges_{layer_name}"
-                #     if process_edge_distance else None
-                # ),
-                # centroid_distance_name=(
-                #     f"Centroid_links_{layer_name}"
-                #     if load_to_canvas else None
-                # ),
-                # edge_distance_name=(
-                #     f"Edge_links_{layer_name}"
-                #     if load_to_canvas else None
-                # ),
-            )
-        else:
-            raise NoUniqueFieldError
-
-    def run_queries(self):
-        self.update_progress()
-        layer_inputs = []
-        load_to_canvas = self.create_distances_files_chb.isChecked()
-        output_dir = str(self.output_dir_le.text())
-        only_selected_features = self.use_selected_features_chb.isChecked()
-
-        kwargs = {}
-        if len(self.model.layers_to_process) > 1:
-            first = self.model.layers_to_process[0]
-            kwargs.update({
-                "default_node_identifier_name": first.id_attribute_field_name,
-                "default_node_attribute_name": first.attribute_field_name,
-                # "default_process_centroid_distance": first.calculate_centroid_distance,
-                # "default_process_edge_distance": first.calculate_edge_distance,
-            })
-
-        for idx, data_item in enumerate(self.model.layers_to_process):
-            if idx == 0 or not self.lock_layers_chb.isChecked():
-                input_parameters = self.get_conefor_input_parameters(
-                    data_item, load_to_canvas)
-            else:
-                input_parameters = self.get_conefor_input_parameters(
-                    data_item, load_to_canvas, **kwargs)
-            layer_inputs.append(input_parameters)
-        self.change_ui_availability(False)
-        # self.processing_task = tasks.LayerProcessorTask(
-        #     description="Generate Conefor input files",
-        #     layers_data=layer_inputs,
-        #     output_dir=Path(output_dir),
-        #     use_selected_features=only_selected_features
-        # )
-        task_manager = qgis.core.QgsApplication.taskManager()
-        for layer_to_process in layer_inputs:
-            common_params = {
-                coneforinputs.ConeforInputsPolygon.INPUT_NODE_IDENTIFIER_NAME[0]: (
-                    layer_to_process.id_attribute_field_name or ""),
-                coneforinputs.ConeforInputsPolygon.INPUT_NODE_ATTRIBUTE_NAME[0]: (
-                    layer_to_process.attribute_field_name or ""),
-                coneforinputs.ConeforInputsPolygon.INPUT_DISTANCE_THRESHOLD[0]: "",
-                coneforinputs.ConeforInputsPolygon.INPUT_OUTPUT_DIRECTORY[0]: str(output_dir),
-            }
-            log(f"{common_params=}")
-            if layer_to_process.layer.geometryType() == qgis.core.Qgis.GeometryType.Polygon:
-                log(f"Creating the conefor:inputsfrompolygon runner task")
-                task = qgis.core.QgsProcessingAlgRunnerTask(
-                    algorithm=self.inputs_from_polygons_algorithm,
-                    parameters={
-                        coneforinputs.ConeforInputsPolygon.INPUT_POLYGON_LAYER[0]: (
-                            layer_to_process.layer),
-                        coneforinputs.ConeforInputsPolygon.INPUT_NODE_CONNECTION_DISTANCE_METHOD[0]: (
-                            layer_to_process.connections_method.value),
-                        **common_params
-                    },
-                    context=self.processing_context
-                )
-            elif layer_to_process.layer.geometryType() == qgis.core.Qgis.GeometryType.Point:
-                task = qgis.core.QgsProcessingAlgRunnerTask(
-                    algorithm=self.inputs_from_points_algorithm,
-                    parameters={
-                        coneforinputs.ConeforInputsPoint.INPUT_POINT_LAYER[0]: (
-                            layer_to_process.layer),
-                        **common_params
-                    },
-                    context=self.processing_context
-                )
-            else:
-                raise RuntimeError(
-                    f"layer: {layer_to_process.layer.name()!r} has invalid "
-                    f"geometry type: {layer_to_process.layer.geometryType()!r}"
-                )
-            self.processing_tasks.append(task)
-            log(f"{self.processing_tasks=}")
-            task_manager.addTask(task)
-
-    def update_progress(self):
-        self.progressBar.setValue(self.processor.global_progress)
-
-    def update_info(self, info, section=0):
-        """
-        Update the progress label with the input info string.
-
-        The information displayed in the progress label is a string composed
-        of three sections:
-
-        * section 0
-        * section 1
-        * section 2
-
-        Inputs:
-
-            info - a string with the information to display in the progress
-                   label
-            section - an integer specifying where in the displayed string
-                      should the 'info' argument be placed.
-        """
-
-        if section == 0:
-            self.progress_la.setText(info)
-        else:
-            current_text = self.progress_la.text()
-            sections = current_text.split(' - ')
-            try:
-                sections[section] = info
-            except IndexError:
-                sections.append(info)
-            self.progress_la.setText(' - '.join(sections))
-        if "ERROR" in info:
-            palette = QtGui.QPalette()
-            palette.setColor(QtGui.QPalette.Foreground, QtCore.Qt.GlobalColor.red)
-            self.progress_la.setPalette(palette)
-
     def toggle_run_button(self):
         """
         Toggle the active state of the run button based on the availability
@@ -414,7 +199,6 @@ class ConeforDialog(QtWidgets.QDialog, FORM_CLASS):
             self.output_la,
             self.output_dir_le,
             self.output_dir_btn,
-            self.progressBar,
             self.buttonBox.button(self.buttonBox.Ok),
         ]
         for widget in widgets:
